cloud functions umorganisiert
