+++ functions/admin.js
@@ -1,0 +1,4 @@
+const admin = require("firebase-admin");
+admin.initializeApp();
+
+exports.admin = admin;
